requirements-prod.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.15.1
python-dotenv==1.1.0
sqlparse==0.5.3
whitenoise>=6.0.0
psycopg2-binary>=2.9.9
requests==2.32.3
Pillow==10.4.0  # Required for image processing
moviepy==1.0.3  # Lighter alternative to OpenCV for video processing
tqdm==4.67.1  # Required for progress tracking 